requirements.txt
apache-beam[gcp]==2.56.0
google-cloud-pubsub==2.21.1
google-cloud-bigquery==3.21.0
google-cloud-storage==2.16.0
google-auth==2.29.0
